fix(read): store scroll_position as a [0,1] fraction to match clients
The mobile/web readers persist reading progress as offsetY/contentHeight
(a fraction in [0,1]) since #257, but the backend was never migrated:
scroll_position stayed an INTEGER column with an *int DTO. Go's JSON
decoder rejects a fractional number like 0.5 into an int, so every
PATCH that carried real progress returned 400 and nothing was saved —
the stored value stayed 0 and articles always reopened at the top.

Align the backend with the fraction model:
- migration 000003: scroll_position INTEGER -> NUMERIC(5,4), CHECK
  tightened to [0,1]; legacy absolute offsets (> 1) are clamped to 0
  before the type change (they overflow NUMERIC(5,4) and clients already
  ignore values > 1 on restore).
- Go types int/*int -> float64/*float64 in the model, DTOs and the
  repository UpdateMetadata signature.
- Reject out-of-range values with a friendly 400 via request validation.
- OpenAPI: scroll_position becomes number(float) bounded [0,1].
- Tests updated to exercise fractional values.

+++ services/read/content/internal/api/handlers/user_content_url_submission_test.go
@@ -114,7 +114,7 @@ func (m *mockUserContentRepo) UpdateWithTx(ctx context.Context, tx *sql.Tx, uc *
 	return nil
 }
 
-func (m *mockUserContentRepo) UpdateMetadata(ctx context.Context, id uuid.UUID, status *string, scrollPosition *int, isFavorite *bool) error {
+func (m *mockUserContentRepo) UpdateMetadata(ctx context.Context, id uuid.UUID, status *string, scrollPosition *float64, isFavorite *bool) error {
 	return nil
 }
